requirements.txt
streamlit
openai
pandas
streamlit_chat
textblob
fpdf
python-dotenv
